userId added in db and delete only user they have added expense

+++ backend/service/expenseService.js
@@ -1,19 +1,23 @@
-const Expense = require("../model/expenseSchema")
+const { Expense } = require("../model/index");
 
-const addExpenseService =async(amount,description,category)=>{
-    const result = await Expense.create({amount,description,category})
-    return result
-}
+const addExpenseService = async (amount, description, category, userId) => {
+  const result = await Expense.create({
+    amount,
+    description,
+    category,
+    userId,
+  });
+  return result;
+};
 
-const getExpenseService =async()=>{
-    const result = await Expense.findAll()
-    return result
-}
+const getExpenseService = async () => {
+  const result = await Expense.findAll();
+  return result;
+};
 
-const deleteExpenseService =async(id)=>{
-    const result = await Expense.destroy({where: {id}})
-    return result
-}
+const deleteExpenseService = async (id,userId ) => {
+  const result = await Expense.destroy({ where: { id, userId } });
+  return result;
+};
 
-
-module.exports ={addExpenseService,getExpenseService, deleteExpenseService }
+module.exports = { addExpenseService, getExpenseService, deleteExpenseService };
